Reorganizar importaciones para utilizar componentes compartidos en las páginas de configuración
